requirements.txt
pandas
numpy
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
# AI Agent
langchain==0.2.16
langchain-ollama==0.1.3
langchain-community==0.2.16
langchain-groq
requests==2.32.3